requirements.txt
torch>=2.0.0
torchvision>=0.15.0
gradio>=5.0.0
huggingface_hub>=0.20.0
Pillow>=9.0.0
